ecommerce_backend copy/requirement.txt
Flask==2.3.5
Flask-Cors==4.0.0
mysql-connector-python==8.1.0
